+++ src/Symfony/Component/Translation/DataCollector/TranslationDataCollector.php
@@ -45,7 +45,7 @@ class TranslationDataCollector extends DataCollector implements LateDataCollecto
     public function collect(Request $request, Response $response, \Throwable $exception = null)
     {
         $this->data['locale'] = $this->translator->getLocale();
-        $this->data['fallback_locales'] = $this->translator->getFallbackLocales();
+        $this->data['fallback_locales'] = $this->translator->getFallbackLocaleProvider()->getUltimateFallbackLocales();
     }
 
     public function reset()