+++ src/Symfony/Component/Translation/LoggingTranslator.php
@@ -76,6 +76,8 @@ class LoggingTranslator implements TranslatorInterface, TranslatorBagInterface, 
      */
     public function getFallbackLocales(): array
     {
+        trigger_deprecation('symfony/translation', '6.2', 'LoggingTranslator::getFallbackLocales() is deprecated. Get the FallbackLocaleProvider instead.');
+
         if ($this->translator instanceof Translator || method_exists($this->translator, 'getFallbackLocales')) {
             return $this->translator->getFallbackLocales();
         }