+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Parameters implementing `TranslatableInterface` are processed
  * Add the file extension to the `XliffFileDumper` constructor
+ * Add `FallbackLocaleProvider` to configure the fallback locale chain in `Translator`
 
 5.4
 ---